# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: 智慧物业管理系统 SaaS 多租户版 API
  version: 1.1.1
  description: |
    本接口契约用于约束 Codex 开发。实际开发不得擅自更改路径、核心字段、状态枚举和权限标识。
    所有业务接口默认返回 ApiResponse；列表接口 data 使用 PageResult。
servers:
  - url: https://api.example.com
    description: 生产环境示例
  - url: http://localhost:8080
    description: 本地开发环境
security:
  - bearerAuth: []
...
paths:
...
  /api/base/vehicles:
    get:
      tags: [Base]
      summary: 车辆列表
      x-permission: base:vehicle:list
      parameters:
        - $ref: '#/components/parameters/PageNo'
        - $ref: '#/components/parameters/PageSize'
        - name: projectId
          in: query
          schema: { type: integer, format: int64 }
        - name: keyword
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
      responses:
        '200': { $ref: '#/components/responses/PageResponse' }
    post:
      tags: [Base]
      summary: 新增车辆
      x-permission: base:vehicle:create
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VehicleRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/base/vehicles/{vehicleId}:
    get:
      tags: [Base]
      summary: 车辆详情
      x-permission: base:vehicle:view
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
    put:
      tags: [Base]
      summary: 修改车辆
      x-permission: base:vehicle:update
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VehicleRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/base/vehicles/{vehicleId}/monthly-rent:
    put:
      tags: [Base]
      summary: 更新车辆月租状态
      x-permission: base:vehicle:monthlyRent
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MonthlyRentRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/base/parking-spaces:
    get:
      tags: [Base]
      summary: 车位列表
      x-permission: base:parkingSpace:list
      parameters:
        - $ref: '#/components/parameters/PageNo'
        - $ref: '#/components/parameters/PageSize'
        - name: projectId
          in: query
          schema: { type: integer, format: int64 }
        - name: keyword
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
      responses:
        '200': { $ref: '#/components/responses/PageResponse' }
    post:
      tags: [Base]
      summary: 新增车位
      x-permission: base:parkingSpace:create
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ParkingSpaceRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/base/parking-spaces/{spaceId}:
    get:
      tags: [Base]
      summary: 车位详情
      x-permission: base:parkingSpace:view
      parameters:
        - name: spaceId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
    put:
      tags: [Base]
      summary: 修改车位
      x-permission: base:parkingSpace:update
      parameters:
        - name: spaceId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ParkingSpaceRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
...
  /api/device/parking:
    get:
      tags: [Device]
      summary: 停车对接记录
      x-permission: device:parking:list
      parameters:
        - $ref: '#/components/parameters/PageNo'
        - $ref: '#/components/parameters/PageSize'
        - name: projectId
          in: query
          schema: { type: integer, format: int64 }
        - name: status
          in: query
          schema: { type: string }
      responses:
        '200': { $ref: '#/components/responses/PageResponse' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageNo:
      name: pageNo
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, default: 20, minimum: 1, maximum: 200 }
...
  responses:
    ApiResponse:
      description: OK
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
    PageResponse:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  data: { $ref: '#/components/schemas/PageResult' }
...
  schemas:
    ApiResponse:
      type: object
      required: [code, message, traceId]
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: success }
        data: { type: object, nullable: true }
        traceId: { type: string }
...
    PageResult:
      type: object
      properties:
        records:
          type: array
          items: { type: object }
        pageNo: { type: integer }
        pageSize: { type: integer }
        total: { type: integer, format: int64 }
...
    ParkingSpaceRequest:
      type: object
      required: [projectId, spaceNo]
      properties:
        projectId: { type: integer, format: int64 }
        spaceNo: { type: string }
        spaceType: { type: string }
        status: { type: string, enum: [AVAILABLE, OCCUPIED, LOCKED, DISABLED] }
        houseId: { type: integer, format: int64, nullable: true }
    VehicleRequest:
      type: object
      required: [projectId, plateNo]
      properties:
        projectId: { type: integer, format: int64 }
        plateNo: { type: string }
        vehicleType: { type: string }
        memberId: { type: integer, format: int64, nullable: true }
        houseId: { type: integer, format: int64, nullable: true }
        spaceId: { type: integer, format: int64, nullable: true }
        monthlyRentStatus: { type: string, enum: [NONE, ACTIVE, EXPIRED, SUSPENDED] }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        status: { type: string, enum: [ACTIVE, DISABLED] }
    MonthlyRentRequest:
      type: object
      required: [monthlyRentStatus]
      properties:
        monthlyRentStatus: { type: string, enum: [NONE, ACTIVE, EXPIRED, SUSPENDED] }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
...